test: improve e2e test performance
