+++ test/helpers/helpers.dart
@@ -1,1 +1,2 @@
 export 'command_helper.dart';
+export 'copy_directory.dart';
